HasRole

+++ app/Http/Controllers/UserController.php
@@ -6,45 +6,58 @@ use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserEditRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
     public function index(){
+
+        abort_if(Gate::denies('user_index'), 403);
+
         $users = User::paginate(3);
         return view('users.index', compact('users'));
     }
 
 
     public function create(){
-        return view('users.create');
+
+        abort_if(Gate::denies('user_create'), 403);
+
+        $roles = Role::all()->pluck('name', 'id');
+        return view('users.create', compact('roles'));
     }
 
 
     public function store(UserCreateRequest $request){
 
-      /*   $request->validate([
-            'name' => 'required|min:3|max:10',
-            'username' => 'required',
-            'email' => 'required|email|unique:users',
-            'password' => 'required'
-        ]);
- */
         $user = User::create($request->only('name', 'username', 'email') + [
             'password' => bcrypt($request->input('password'))
         ]);
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'usuario creado correctamente');
     }
 
 
     public function show(User $user){
+
+        abort_if(Gate::denies('user_show'), 403);
         //$user = User::findOrFail($id);
+        $user->load('roles');
         return view('users.show', compact('user'));
     }
 
 
     public function edit(User $user){
+
+        abort_if(Gate::denies('user_edit'), 403);
+
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
         
-        return view('users.edit', compact('user'));
+        return view('users.edit', compact('user', 'roles'));
     }
 
 
@@ -56,11 +69,20 @@ class UserController extends Controller
             $data['password'] = bcrypt($password);
         
         $user->update($data);
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario actualizado satisfactoriamente');
     }
 
 
     public function destroy(User $user){
+
+        abort_if(Gate::denies('user_destroy'), 403);
+
+        if (auth()->user()->id == $user->id) {
+            return redirect()->route('users.index');
+        }
+
         $user->delete();
         return back()->with('success', 'Usuario eliminado correctamente');
     }
